cleanup components on staking and withdrawal
