+++ packages/nextjs/components/Header.tsx
@@ -4,7 +4,7 @@ import React, { useRef } from "react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import { hardhat } from "viem/chains";
-import { Bars3Icon, BugAntIcon } from "@heroicons/react/24/outline";
+import { Bars3Icon, BugAntIcon, UserGroupIcon } from "@heroicons/react/24/outline";
 import { CircleStackIcon, InboxStackIcon } from "@heroicons/react/24/outline";
 import { FaucetButton, RainbowKitCustomConnectButton } from "~~/components/scaffold-eth";
 import { useOutsideClick, useTargetNetwork } from "~~/hooks/scaffold-eth";
@@ -20,6 +20,11 @@ export const menuLinks: HeaderMenuLink[] = [
     label: "Stake",
     href: "/stakeandreturn",
     icon: <CircleStackIcon className="h-4 w-4" />,
+  },
+  {
+    label: "Participants",
+    href: "/participants",
+    icon: <UserGroupIcon className="h-4 w-4" />,
   },
   {
     label: "Stake Events",
