+++ packages/nextjs/utils/scaffold-eth/decodeTxData.ts
@@ -3,6 +3,7 @@ import { GenericContractsDeclaration } from "./contract";
 import { Abi, AbiFunction, decodeFunctionData, getAbiItem } from "viem";
 import { hardhat } from "viem/chains";
 import contractData from "~~/contracts/deployedContracts";
+import { IGetWithdrawEstimate } from "~~/types/abitype/interfaces";
 
 type ContractsInterfaces = Record<string, Abi>;
 type TransactionType = TransactionWithFunction | null;
@@ -65,3 +66,21 @@ export const getFunctionDetails = (transaction: TransactionType) => {
 };
 
 export const weiToEth = (wei: bigint) => Number(wei) / 1e18;
+
+export const parseGetWithdrawEstimate = (res: readonly [bigint, bigint, bigint] | undefined): IGetWithdrawEstimate => {
+  if (!res || res.length < 2) {
+    return { stake: BigInt(-1), reward: BigInt(0), total: BigInt(0) };
+  }
+  return {
+    stake: res[0] ?? BigInt(0),
+    reward: res[1] ?? BigInt(0),
+    total: res[2] ?? BigInt(0),
+  };
+};
+
+export const getPercentInterestIncrease = (stake: bigint, reward: bigint, total: bigint): string => {
+  if (!stake || !reward || !total) return "-1"; // prevent div by zero
+  const ethTotal = weiToEth(total);
+  const ethStake = weiToEth(stake);
+  return Number(((ethTotal - ethStake) / ethStake) * 99).toFixed(6);
+};
